Create quotesFavoris component
